create summary file.
